moved themeOptions to styles folder
