+++ src/App.js
@@ -13,7 +13,7 @@ import Signup from "./pages/Signup"
 import PetBio from "./pages/PetBio"
 import Favorites from "./pages/Favorites"
 import { UserContext, PetsContext, SearchContext } from "./context"
-import { themeOptions } from "./themeOptions"
+import { themeOptions } from "./Styles/themeOptions"
 import { ThemeProvider } from "@mui/material/styles"
 function App() {
   //login state can be used for conditional redering to swap login link for user page link or favorites page link and userProfile info can be used to populate infor with other forms
@@ -41,40 +41,40 @@ function App() {
 
   return (
     <div className="App">
-      <ThemeProvider theme= {themeOptions}>
-      <UserContext.Provider
-        value={{ isLoggedIn, setIsLoggedIn, userProfile, setUserProfile }}
-      >
-        <SearchContext.Provider
-          value={{
-            search,
-            setSearch,
-            handleSearch,
-            pageNumber,
-            setPageNumber,
-            pageCount,
-            setPageCount,
-          }}
+      <ThemeProvider theme={themeOptions}>
+        <UserContext.Provider
+          value={{ isLoggedIn, setIsLoggedIn, userProfile, setUserProfile }}
         >
-          <PetsContext.Provider
-            value={{ petList, setPetList, isFavoritedList, setIsFavoritedList }}
+          <SearchContext.Provider
+            value={{
+              search,
+              setSearch,
+              handleSearch,
+              pageNumber,
+              setPageNumber,
+              pageCount,
+              setPageCount,
+            }}
           >
-            <Navbar />
-            <Routes>
-              <Route path="/" element={<Home />} />
-              <Route path="adopt" element={<Adopt />} />
-              <Route path="volunteer" element={<Volunteer />} />
-              <Route path="donate" element={<Donate />} />
-              <Route path="about" element={<About />} />
-              <Route path="login" element={<Login />} />
-              <Route path="/petbio/:id" element={<PetBio />} />
-              <Route path="Signup" element={<Signup />} />
-              <Route path="favorites" element={<Favorites />} />
-            </Routes>
-            <Footer />
-          </PetsContext.Provider>
-        </SearchContext.Provider>
-      </UserContext.Provider>
+            <PetsContext.Provider
+              value={{ petList, setPetList, isFavoritedList, setIsFavoritedList }}
+            >
+              <Navbar />
+              <Routes>
+                <Route path="/" element={<Home />} />
+                <Route path="adopt" element={<Adopt />} />
+                <Route path="volunteer" element={<Volunteer />} />
+                <Route path="donate" element={<Donate />} />
+                <Route path="about" element={<About />} />
+                <Route path="login" element={<Login />} />
+                <Route path="/petbio/:id" element={<PetBio />} />
+                <Route path="Signup" element={<Signup />} />
+                <Route path="favorites" element={<Favorites />} />
+              </Routes>
+              <Footer />
+            </PetsContext.Provider>
+          </SearchContext.Provider>
+        </UserContext.Provider>
       </ThemeProvider>
     </div>
   )
